requirements.txt
python-snap7==2.0.2
numpy==1.19.5
pytango==9.3.2